moves Encodable Decodable to types.rs
