+++ src/types.rs
@@ -25,6 +25,7 @@ use rand::random;
 use sodiumoxide;
 use NameType;
 use std::fmt;
+use RoutingError;
 
 pub fn array_as_vector(arr: &[u8]) -> Vec<u8> {
   let mut vector = Vec::new();
@@ -432,6 +433,35 @@ pub enum MessageTypeTag {
   AccountTransfer
 }
 
+impl Encodable for RoutingError {
+    fn encode<E: Encoder>(&self, e: &mut E)->Result<(), E::Error> {
+        let mut type_tag;
+        match *self {
+            RoutingError::Success => type_tag = "Success",
+            RoutingError::FailedToBootstrap => type_tag = "FailedToBootstrap",
+            RoutingError::NoData => type_tag = "NoData",
+            RoutingError::InvalidRequest => type_tag = "InvalidRequest",
+            RoutingError::IncorrectData(_) => type_tag = "IncorrectData",
+        };
+        CborTagEncode::new(5483_100, &(&type_tag)).encode(e)
+    }
+}
+
+impl Decodable for RoutingError {
+    fn decode<D: Decoder>(d: &mut D)->Result<RoutingError, D::Error> {
+        try!(d.read_u64());
+        let mut type_tag : String;
+        type_tag = try!(Decodable::decode(d));
+        match &type_tag[..] {
+            "Success" => Ok(RoutingError::Success),
+            "FailedToBootstrap" => Ok(RoutingError::FailedToBootstrap),
+            "NoData" => Ok(RoutingError::NoData),
+            "InvalidRequest" => Ok(RoutingError::InvalidRequest),
+            "IncorrectData" => Ok(RoutingError::IncorrectData(vec![])),
+            _ => Ok(RoutingError::IncorrectData("unknown".to_string().into_bytes()))
+        }
+    }
+}
 
 #[cfg(test)]
 #[allow(deprecated)]
